chore: update App.js
