+++ src/App.js
@@ -1,126 +1,21 @@
-import React, { PureComponent } from 'react'
-import PropTypes from 'prop-types'
-import { withRouter } from 'react-router-dom'
-import { Link } from 'react-router-dom'
-import Sidebar from 'd2-ui/lib/sidebar/Sidebar.component'
-import FeedbackSnackbar from 'd2-ui/lib/feedback-snackbar/FeedbackSnackbar.component'
-import { LOADING } from 'd2-ui/lib/feedback-snackbar/FeedbackSnackbarTypes'
-import CircularProgress from 'd2-ui/lib/circular-progress/CircularProgress'
-import i18n from './locales'
-import { injectTranslationsToD2 } from './configI18n'
-import AppRouter from './components/app-router/AppRouter'
-import { sections } from './pages/sections.conf'
-import './custom-css/D2UISidebarOverrides.css'
-import { HeaderBar } from '@dhis2/ui-widgets'
+import { useD2 } from '@dhis2/app-runtime-adapter-d2'
+import React from 'react'
+import { HashRouter } from 'react-router-dom'
+import App from './components/App'
+import './locales'
 
-import styles from './App.module.css'
+const AppWrapper = () => {
+    const { d2 } = useD2()
 
-export class DumbApp extends PureComponent {
-    static childContextTypes = {
-        showSnackbar: PropTypes.bool,
-        snackbarConf: PropTypes.shape({
-            type: PropTypes.string,
-            message: PropTypes.string,
-            duration: PropTypes.number,
-        }),
-        currentSection: PropTypes.string,
-        updateAppState: PropTypes.func,
+    if (!d2) {
+        return null
     }
 
-    static contextTypes = {
-        d2: PropTypes.object,
-    }
-
-    constructor(props, context) {
-        super(props)
-
-        injectTranslationsToD2(context.d2)
-
-        this.state = {
-            currentSection: '',
-            showSnackbar: false,
-            snackbarConf: {
-                type: '',
-                message: '',
-            },
-        }
-
-        this.updateAppState = this.updateAppState.bind(this)
-    }
-
-    getChildContext() {
-        return {
-            showSnackbar: this.state.showSnackbar,
-            snackbarConf: this.state.snackbarConf,
-            currentSection: this.state.currentSection,
-            updateAppState: this.updateAppState,
-        }
-    }
-
-    updateAppState(appState) {
-        if (
-            appState.currentSection &&
-            !appState.pageState &&
-            this.state.currentSection !== appState.currentSection
-        ) {
-            // clear page state because we are updating page
-            this.setState({
-                ...appState,
-                pageState: undefined,
-                showSnackbar: false,
-            })
-        } else {
-            this.setState(appState)
-        }
-    }
-
-    showLoadingProgress() {
-        // FIXME: remove loading snackbar info and replace with loading state
-        return (
-            this.state.snackbarConf.type === LOADING && this.state.showSnackbar
-        )
-    }
-
-    render() {
-        const nonOnChangeSection = () => null
-        const translatedSections = sections.map(section =>
-            Object.assign(section, {
-                icon: section.info.icon,
-                label: i18n.t(section.info.label),
-                containerElement: <Link to={section.path} />,
-            })
-        )
-
-        const feedbackElement = this.showLoadingProgress() ? (
-            <div id={'circularLoadingId'} className={styles.centered}>
-                <CircularProgress />
-            </div>
-        ) : (
-            <span id={'feedbackSnackbarId'}>
-                <FeedbackSnackbar
-                    show={this.state.showSnackbar}
-                    conf={this.state.snackbarConf}
-                />
-            </span>
-        )
-
-        return (
-            <div>
-                <HeaderBar appName={i18n.t('Data Administration')} />
-                <Sidebar
-                    sections={translatedSections}
-                    currentSection={this.state.currentSection}
-                    onChangeSection={nonOnChangeSection}
-                />
-                <div className={styles.contentWrapper}>
-                    <div className={styles.contentArea}>
-                        <AppRouter pageState={this.state.pageState} />
-                    </div>
-                </div>
-                {feedbackElement}
-            </div>
-        )
-    }
+    return (
+        <HashRouter>
+            <App d2={d2} />
+        </HashRouter>
+    )
 }
 
-export default withRouter(DumbApp)
+export default AppWrapper
