--- src/index.css
@@ -1,41 +1,0 @@
-html {
-    background: #f3f3f3;
-    font-family: 'Roboto', sans-serif;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-}
-
-h1 {
-    font-family: Roboto;
-    font-size: 24px;
-    font-weight: 300;
-    letter-spacing: 1.2px;
-    color: rgba(0, 0, 0, 0.87);
-}
-
-.material-icons {
-    font-family: 'Material Icons';
-    font-weight: normal;
-    font-style: normal;
-    font-size: 24px; /* Preferred icon size */
-    display: inline-block;
-    line-height: 1;
-    text-transform: none;
-    letter-spacing: normal;
-    word-wrap: normal;
-    white-space: nowrap;
-    direction: ltr;
-    /* Support for all WebKit browsers. */
-    -webkit-font-smoothing: antialiased;
-    /* Support for Safari and Chrome. */
-    text-rendering: optimizeLegibility;
-    /* Support for Firefox. */
-    -moz-osx-font-smoothing: grayscale;
-    /* Support for IE. */
-    -webkit-font-feature-settings: 'liga';
-    font-feature-settings: 'liga';
-    font-display: block;
-}
